Move surface element to seperate file
