+++ src/openrct2/world/Surface.h
@@ -1,0 +1,120 @@
+#pragma region Copyright (c) 2014-2017 OpenRCT2 Developers
+/*****************************************************************************
+* OpenRCT2, an open source clone of Roller Coaster Tycoon 2.
+*
+* OpenRCT2 is the work of many authors, a full list can be found in contributors.md
+* For more information, visit https://github.com/OpenRCT2/OpenRCT2
+*
+* OpenRCT2 is free software: you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation, either version 3 of the License, or
+* (at your option) any later version.
+*
+* A full copy of the GNU General Public License can be found in licence.txt
+*****************************************************************************/
+#pragma endregion
+
+#pragma once
+
+struct rct_tile_element;
+
+enum {
+    TERRAIN_GRASS,
+    TERRAIN_SAND,
+    TERRAIN_DIRT,
+    TERRAIN_ROCK,
+    TERRAIN_MARTIAN,
+    TERRAIN_CHECKERBOARD,
+    TERRAIN_GRASS_CLUMPS,
+    TERRAIN_ICE,
+    TERRAIN_GRID_RED,
+    TERRAIN_GRID_YELLOW,
+    TERRAIN_GRID_BLUE,
+    TERRAIN_GRID_GREEN,
+    TERRAIN_SAND_DARK,
+    TERRAIN_SAND_LIGHT,
+    TERRAIN_COUNT_REGULAR = 14, // The amount of surface types the user can actually select - what follows are technical types
+    TERRAIN_CHECKERBOARD_INVERTED = 14,
+    TERRAIN_UNDERGROUND_VIEW,
+};
+
+enum {
+    TERRAIN_EDGE_ROCK,
+    TERRAIN_EDGE_WOOD_RED,
+    TERRAIN_EDGE_WOOD_BLACK,
+    TERRAIN_EDGE_ICE,
+
+    TERRAIN_EDGE_RCT2_COUNT,
+
+    TERRAIN_EDGE_BRICK = TERRAIN_EDGE_RCT2_COUNT,
+    TERRAIN_EDGE_IRON,
+    TERRAIN_EDGE_GREY,
+    TERRAIN_EDGE_YELLOW,
+    TERRAIN_EDGE_RED,
+    TERRAIN_EDGE_PURPLE,
+    TERRAIN_EDGE_GREEN,
+    TERRAIN_EDGE_STONE_BROWN,
+    TERRAIN_EDGE_STONE_GREY,
+    TERRAIN_EDGE_SKYSCRAPER_A,
+    TERRAIN_EDGE_SKYSCRAPER_B,
+
+    TERRAIN_EDGE_COUNT
+};
+
+enum {
+    GRASS_LENGTH_MOWED,
+    GRASS_LENGTH_CLEAR_0,
+    GRASS_LENGTH_CLEAR_1,
+    GRASS_LENGTH_CLEAR_2,
+    GRASS_LENGTH_CLUMPS_0,
+    GRASS_LENGTH_CLUMPS_1,
+    GRASS_LENGTH_CLUMPS_2
+};
+
+enum {
+    OWNERSHIP_UNOWNED = 0,
+    OWNERSHIP_CONSTRUCTION_RIGHTS_OWNED = (1 << 4),
+    OWNERSHIP_OWNED = (1 << 5),
+    OWNERSHIP_CONSTRUCTION_RIGHTS_AVAILABLE = (1 << 6),
+    OWNERSHIP_AVAILABLE = (1 << 7)
+};
+
+enum {
+    TILE_ELEMENT_SLOPE_FLAT = 0x00,
+    TILE_ELEMENT_SLOPE_ALL_CORNERS_UP = 0x0F,
+
+    TILE_ELEMENT_SLOPE_N_CORNER_UP = (1 << 0),
+    TILE_ELEMENT_SLOPE_E_CORNER_UP = (1 << 1),
+    TILE_ELEMENT_SLOPE_S_CORNER_UP = (1 << 2),
+    TILE_ELEMENT_SLOPE_W_CORNER_UP = (1 << 3),
+    TILE_ELEMENT_SLOPE_DOUBLE_HEIGHT = (1 << 4),
+
+    TILE_ELEMENT_SLOPE_W_CORNER_DN = TILE_ELEMENT_SLOPE_ALL_CORNERS_UP & ~TILE_ELEMENT_SLOPE_W_CORNER_UP,
+    TILE_ELEMENT_SLOPE_S_CORNER_DN = TILE_ELEMENT_SLOPE_ALL_CORNERS_UP & ~TILE_ELEMENT_SLOPE_S_CORNER_UP,
+    TILE_ELEMENT_SLOPE_E_CORNER_DN = TILE_ELEMENT_SLOPE_ALL_CORNERS_UP & ~TILE_ELEMENT_SLOPE_E_CORNER_UP,
+    TILE_ELEMENT_SLOPE_N_CORNER_DN = TILE_ELEMENT_SLOPE_ALL_CORNERS_UP & ~TILE_ELEMENT_SLOPE_N_CORNER_UP,
+
+    TILE_ELEMENT_SLOPE_NE_SIDE_UP = TILE_ELEMENT_SLOPE_N_CORNER_UP | TILE_ELEMENT_SLOPE_E_CORNER_UP,
+    TILE_ELEMENT_SLOPE_SE_SIDE_UP = TILE_ELEMENT_SLOPE_E_CORNER_UP | TILE_ELEMENT_SLOPE_S_CORNER_UP,
+    TILE_ELEMENT_SLOPE_NW_SIDE_UP = TILE_ELEMENT_SLOPE_N_CORNER_UP | TILE_ELEMENT_SLOPE_W_CORNER_UP,
+    TILE_ELEMENT_SLOPE_SW_SIDE_UP = TILE_ELEMENT_SLOPE_S_CORNER_UP | TILE_ELEMENT_SLOPE_W_CORNER_UP,
+
+    TILE_ELEMENT_SLOPE_W_E_VALLEY = TILE_ELEMENT_SLOPE_E_CORNER_UP | TILE_ELEMENT_SLOPE_W_CORNER_UP,
+    TILE_ELEMENT_SLOPE_N_S_VALLEY = TILE_ELEMENT_SLOPE_N_CORNER_UP | TILE_ELEMENT_SLOPE_S_CORNER_UP
+};
+
+// Surface
+#define TILE_ELEMENT_SURFACE_DIAGONAL_FLAG       0x10 // in rct_tile_element.properties.surface.slope
+#define TILE_ELEMENT_SURFACE_RAISED_CORNERS_MASK 0x0F // in rct_tile_element.properties.surface.slope
+#define TILE_ELEMENT_SURFACE_SLOPE_MASK          (TILE_ELEMENT_SURFACE_DIAGONAL_FLAG | TILE_ELEMENT_SURFACE_RAISED_CORNERS_MASK) // in rct_tile_element.properties.surface.slope
+#define TILE_ELEMENT_SURFACE_EDGE_STYLE_MASK     0xE0 // in rct_tile_element.properties.surface.slope
+#define TILE_ELEMENT_SURFACE_WATER_HEIGHT_MASK   0x1F // in rct_tile_element.properties.surface.terrain
+#define TILE_ELEMENT_SURFACE_TERRAIN_MASK        0xE0 // in rct_tile_element.properties.surface.terrain
+
+sint32 tile_element_get_terrain(const rct_tile_element *element);
+sint32 tile_element_get_terrain_edge(const rct_tile_element *element);
+void tile_element_set_terrain(rct_tile_element *element, sint32 terrain);
+void tile_element_set_terrain_edge(rct_tile_element *element, sint32 terrain);
+
+// ~Oli414: Needs to renamed. This function is specific to the surface object.
+sint32 map_get_water_height(const rct_tile_element * tileElement);
